CSVGen relocation attempt 2
