+++ CheckInTerminal/src/part2/ManagementGUI.java
@@ -1,4 +1,4 @@
-package CheckInTerminal;
+package part2;
 import javax.swing.*;
 import javax.swing.border.*;
 import java.awt.*;
